# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "绦管后台 API 文档",
    "description": "电商后台管理系统RESTful API，支持小程序对接",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://qimengzhiyue.cn/api/v1",
      "description": "生产环境"
    },
    {
      "url": "http://localhost:3000/api/v1",
      "description": "开发环境"
    }
  ],
...
  "paths": {
...
    "/products": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "商品列表",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "category_id",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "刕efa商品",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  },
                  "stock": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "刕efa成功"
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "商品详情",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          },
          "404": {
            "description": "不存在"
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "summary": "更新商品",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "summary": "删除商品",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          }
        }
      }
    },
...
    "/products/search": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "搜索商品",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "搜索成功"
          }
        }
      }
    },
...
    "/search/suggestions": {
      "get": {
        "tags": [
          "Search"
        ],
        "summary": "搜索建议",
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer Token 认证"
      }
    },
...
  }
}